requirements.txt
# ============================================================
#  requirements.txt
#  Instale com: pip install -r requirements.txt
# ============================================================

# HTTP e parsing de RSS
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
lxml_html_clean>=0.1.0

# Playwright — scraping com navegador real (suporta JavaScript/SPAs)
playwright>=1.44.0
# Após instalar, execute: playwright install chromium

# Trafilatura — extração de conteúdo limpo de artigos
trafilatura>=1.8.0

# Variáveis de ambiente
python-dotenv>=1.0.0
